update: rename the classname.js
